--- dd-java-agent/instrumentation/jms-2/jms-2.gradle
@@ -1,34 +1,0 @@
-apply plugin: 'version-scan'
-
-versionScan {
-  group = "javax.jms"
-  module = "javax.jms-api"
-  versions = "[2,)"
-  legacyModule = "jms-api"
-  verifyPresent = [
-    "javax.jms.JMSContext"        : null,
-    "javax.jms.CompletionListener": null,
-  ]
-}
-
-apply from: "${rootDir}/gradle/java.gradle"
-
-dependencies {
-  // use jms1 helpers
-  compile(project(':dd-java-agent:instrumentation:jms-1')) {
-    transitive = false
-  }
-  compileOnly group: 'javax.jms', name: 'javax.jms-api', version: '2.0.1'
-
-  compile deps.bytebuddy
-  compile deps.opentracing
-  annotationProcessor deps.autoservice
-  implementation deps.autoservice
-
-  compile project(':dd-java-agent:agent-tooling')
-
-  testCompile project(':dd-java-agent:testing')
-  testCompile group: 'org.hornetq', name: 'hornetq-jms-client', version: '2.4.7.Final'
-  testCompile group: 'org.hornetq', name: 'hornetq-jms-server', version: '2.4.7.Final'
-
-}
