+++ dd-java-agent/instrumentation/jms/jms.gradle
@@ -11,7 +11,27 @@ versionScan {
   ]
 }
 
+muzzle {
+  pass {
+    group = "javax.jms"
+    module = "jms-api"
+    versions = "(,)"
+  }
+  pass {
+    group = "javax.jms"
+    module = "javax.jms-api"
+    versions = "(,)"
+  }
+}
+
 apply from: "${rootDir}/gradle/java.gradle"
+
+apply plugin: 'org.unbroken-dome.test-sets'
+
+testSets {
+  latestDepTest {
+  }
+}
 
 dependencies {
   compileOnly group: 'javax.jms', name: 'jms-api', version: '1.1-rev-1'
@@ -27,4 +47,7 @@ dependencies {
   testCompile group: 'org.apache.activemq.tooling', name: 'activemq-junit', version: '5.14.5'
   testCompile group: 'org.apache.activemq', name: 'activemq-pool', version: '5.14.5'
   testCompile group: 'org.apache.activemq', name: 'activemq-broker', version: '5.14.5'
+
+  latestDepTestCompile group: 'org.hornetq', name: 'hornetq-jms-client', version: '2.4.7.Final'
+  latestDepTestCompile group: 'org.hornetq', name: 'hornetq-jms-server', version: '2.4.7.Final'
 }
