+++ dd-java-agent/instrumentation/jms/src/main/java/datadog/trace/instrumentation/jms/MessagePropertyTextMap.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.jms.util;
+package datadog.trace.instrumentation.jms;
 
 import io.opentracing.propagation.TextMap;
 import java.util.Enumeration;
@@ -7,7 +7,9 @@ import java.util.Iterator;
 import java.util.Map;
 import javax.jms.JMSException;
 import javax.jms.Message;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 public class MessagePropertyTextMap implements TextMap {
   static final String DASH = "__dash__";
 
@@ -39,10 +41,13 @@ public class MessagePropertyTextMap implements TextMap {
 
   @Override
   public void put(final String key, final String value) {
+    final String propName = key.replace("-", DASH);
     try {
-      message.setStringProperty(key.replace("-", DASH), value);
+      message.setStringProperty(propName, value);
     } catch (final JMSException e) {
-      throw new RuntimeException(e);
+      if (log.isDebugEnabled()) {
+        log.debug("Failure setting jms property: " + propName, e);
+      }
     }
   }
 }
