+++ dd-java-agent/instrumentation/jms/src/test/groovy/JMS1Test.groovy
@@ -1,24 +1,16 @@
-import com.google.common.io.Files
 import datadog.trace.agent.test.AgentTestRunner
 import datadog.trace.agent.test.asserts.ListWriterAssert
 import datadog.trace.api.DDSpanTypes
 import datadog.trace.api.DDTags
 import io.opentracing.tag.Tags
-import org.hornetq.api.core.TransportConfiguration
-import org.hornetq.api.core.client.HornetQClient
-import org.hornetq.api.jms.HornetQJMSClient
-import org.hornetq.api.jms.JMSFactoryType
-import org.hornetq.core.config.Configuration
-import org.hornetq.core.config.CoreQueueConfiguration
-import org.hornetq.core.config.impl.ConfigurationImpl
-import org.hornetq.core.remoting.impl.invm.InVMAcceptorFactory
-import org.hornetq.core.remoting.impl.invm.InVMConnectorFactory
-import org.hornetq.core.remoting.impl.netty.NettyAcceptorFactory
-import org.hornetq.core.server.HornetQServers
-import org.hornetq.jms.client.HornetQMessageConsumer
-import org.hornetq.jms.client.HornetQMessageProducer
+import org.apache.activemq.ActiveMQConnectionFactory
+import org.apache.activemq.ActiveMQMessageConsumer
+import org.apache.activemq.ActiveMQMessageProducer
+import org.apache.activemq.command.ActiveMQTextMessage
+import org.apache.activemq.junit.EmbeddedActiveMQBroker
 import spock.lang.Shared
 
+import javax.jms.Connection
 import javax.jms.Message
 import javax.jms.MessageListener
 import javax.jms.Session
@@ -28,52 +20,28 @@ import java.util.concurrent.atomic.AtomicReference
 
 import static datadog.trace.agent.test.asserts.ListWriterAssert.assertTraces
 
-class JMS2Test extends AgentTestRunner {
+class JMS1Test extends AgentTestRunner {
   @Shared
   String messageText = "a message"
   @Shared
-  static Session session
+  Session session
+
+  ActiveMQTextMessage message = session.createTextMessage(messageText)
 
   def setupSpec() {
-    def tempDir = Files.createTempDir()
-    tempDir.deleteOnExit()
+    EmbeddedActiveMQBroker broker = new EmbeddedActiveMQBroker()
+    broker.start()
+    final ActiveMQConnectionFactory connectionFactory = broker.createConnectionFactory()
 
-    Configuration config = new ConfigurationImpl()
-    config.bindingsDirectory = tempDir.path
-    config.journalDirectory = tempDir.path
-    config.createBindingsDir = false
-    config.createJournalDir = false
-    config.securityEnabled = false
-    config.persistenceEnabled = false
-    config.setQueueConfigurations([new CoreQueueConfiguration("someQueue", "someQueue", null, true)])
-    config.setAcceptorConfigurations([new TransportConfiguration(NettyAcceptorFactory.name),
-                                      new TransportConfiguration(InVMAcceptorFactory.name)].toSet())
-
-    HornetQServers.newHornetQServer(config).start()
-
-    def serverLocator = HornetQClient.createServerLocatorWithoutHA(new TransportConfiguration(InVMConnectorFactory.name))
-    def sf = serverLocator.createSessionFactory()
-    def clientSession = sf.createSession(false, false, false)
-    clientSession.createQueue("jms.queue.someQueue", "jms.queue.someQueue", true)
-    clientSession.createQueue("jms.topic.someTopic", "jms.topic.someTopic", true)
-    clientSession.close()
-    sf.close()
-    serverLocator.close()
-
-    def connectionFactory = HornetQJMSClient.createConnectionFactoryWithoutHA(JMSFactoryType.CF,
-      new TransportConfiguration(InVMConnectorFactory.name))
-
-    def connection = connectionFactory.createConnection()
+    final Connection connection = connectionFactory.createConnection()
     connection.start()
     session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE)
-    session.run()
   }
 
   def "sending a message to #jmsResourceName generates spans"() {
     setup:
     def producer = session.createProducer(destination)
     def consumer = session.createConsumer(destination)
-    def message = session.createTextMessage(messageText)
 
     producer.send(message)
 
@@ -85,7 +53,7 @@ class JMS2Test extends AgentTestRunner {
       producerTrace(it, 0, jmsResourceName)
       trace(1, 1) { // Consumer trace
         span(0) {
-          childOf TEST_WRITER.firstTrace().get(0)
+          childOf TEST_WRITER.firstTrace().get(2)
           serviceName "jms"
           operationName "jms.consume"
           resourceName "Consumed from $jmsResourceName"
@@ -95,9 +63,9 @@ class JMS2Test extends AgentTestRunner {
           tags {
             defaultTags()
             "${DDTags.SPAN_TYPE}" DDSpanTypes.MESSAGE_CONSUMER
-            "${Tags.COMPONENT.key}" "jms2"
+            "${Tags.COMPONENT.key}" "jms"
             "${Tags.SPAN_KIND.key}" "consumer"
-            "span.origin.type" HornetQMessageConsumer.name
+            "span.origin.type" ActiveMQMessageConsumer.name
           }
         }
       }
@@ -129,7 +97,6 @@ class JMS2Test extends AgentTestRunner {
       }
     }
 
-    def message = session.createTextMessage(messageText)
     producer.send(message)
     lock.countDown()
 
@@ -138,7 +105,7 @@ class JMS2Test extends AgentTestRunner {
       producerTrace(it, 0, jmsResourceName)
       trace(1, 1) { // Consumer trace
         span(0) {
-          childOf TEST_WRITER.firstTrace().get(0)
+          childOf TEST_WRITER.firstTrace().get(2)
           serviceName "jms"
           operationName "jms.onMessage"
           resourceName "Received from $jmsResourceName"
@@ -148,9 +115,9 @@ class JMS2Test extends AgentTestRunner {
           tags {
             defaultTags()
             "${DDTags.SPAN_TYPE}" DDSpanTypes.MESSAGE_CONSUMER
-            "${Tags.COMPONENT.key}" "jms2"
+            "${Tags.COMPONENT.key}" "jms"
             "${Tags.SPAN_KIND.key}" "consumer"
-            "span.origin.type" { t -> t.contains("JMS2Test") }
+            "span.origin.type" { t -> t.contains("JMS1Test") }
           }
         }
       }
@@ -192,9 +159,9 @@ class JMS2Test extends AgentTestRunner {
           tags {
             defaultTags()
             "${DDTags.SPAN_TYPE}" DDSpanTypes.MESSAGE_CONSUMER
-            "${Tags.COMPONENT.key}" "jms2"
+            "${Tags.COMPONENT.key}" "jms"
             "${Tags.SPAN_KIND.key}" "consumer"
-            "span.origin.type" HornetQMessageConsumer.name
+            "span.origin.type" ActiveMQMessageConsumer.name
           }
         }
       }
@@ -231,9 +198,9 @@ class JMS2Test extends AgentTestRunner {
           tags {
             defaultTags()
             "${DDTags.SPAN_TYPE}" DDSpanTypes.MESSAGE_CONSUMER
-            "${Tags.COMPONENT.key}" "jms2"
+            "${Tags.COMPONENT.key}" "jms"
             "${Tags.SPAN_KIND.key}" "consumer"
-            "span.origin.type" HornetQMessageConsumer.name
+            "span.origin.type" ActiveMQMessageConsumer.name
           }
         }
       }
@@ -248,8 +215,63 @@ class JMS2Test extends AgentTestRunner {
     session.createTopic("someTopic") | "Topic someTopic"
   }
 
+  def "sending a read-only message to #jmsResourceName fails"() {
+    setup:
+    def producer = session.createProducer(destination)
+    def consumer = session.createConsumer(destination)
+
+    expect:
+    !message.isReadOnlyProperties()
+
+    when:
+    message.setReadOnlyProperties(true)
+    and:
+    producer.send(message)
+
+    TextMessage receivedMessage = consumer.receive()
+
+    then:
+    receivedMessage.text == messageText
+
+    // This will result in a logged failure because we tried to
+    // write properties in MessagePropertyTextMap when readOnlyProperties = true.
+    // The consumer span will also not be linked to the parent.
+    assertTraces(TEST_WRITER, 2) {
+      producerTrace(it, 0, jmsResourceName)
+      trace(1, 1) { // Consumer trace
+        span(0) {
+          parent()
+          serviceName "jms"
+          operationName "jms.consume"
+          resourceName "Consumed from $jmsResourceName"
+          spanType DDSpanTypes.MESSAGE_PRODUCER
+          errored false
+
+          tags {
+            defaultTags()
+            "${DDTags.SPAN_TYPE}" DDSpanTypes.MESSAGE_CONSUMER
+            "${Tags.COMPONENT.key}" "jms"
+            "${Tags.SPAN_KIND.key}" "consumer"
+            "span.origin.type" ActiveMQMessageConsumer.name
+          }
+        }
+      }
+    }
+
+    cleanup:
+    producer.close()
+    consumer.close()
+
+    where:
+    destination                      | jmsResourceName
+    session.createQueue("someQueue") | "Queue someQueue"
+    session.createTopic("someTopic") | "Topic someTopic"
+    session.createTemporaryQueue()   | "Temporary Queue"
+    session.createTemporaryTopic()   | "Temporary Topic"
+  }
+
   def producerTrace(ListWriterAssert writer, int index, String jmsResourceName) {
-    writer.trace(index, 1) {
+    writer.trace(index, 3) {
       span(0) {
         parent()
         serviceName "jms"
@@ -261,9 +283,41 @@ class JMS2Test extends AgentTestRunner {
         tags {
           defaultTags()
           "${DDTags.SPAN_TYPE}" DDSpanTypes.MESSAGE_PRODUCER
-          "${Tags.COMPONENT.key}" "jms2"
+          "${Tags.COMPONENT.key}" "jms"
           "${Tags.SPAN_KIND.key}" "producer"
-          "span.origin.type" HornetQMessageProducer.name
+          "span.origin.type" ActiveMQMessageProducer.name
+        }
+      }
+      span(1) {
+        childOf span(0)
+        serviceName "jms"
+        operationName "jms.produce"
+        resourceName "Produced for $jmsResourceName"
+        spanType DDSpanTypes.MESSAGE_PRODUCER
+        errored false
+
+        tags {
+          defaultTags()
+          "${DDTags.SPAN_TYPE}" DDSpanTypes.MESSAGE_PRODUCER
+          "${Tags.COMPONENT.key}" "jms"
+          "${Tags.SPAN_KIND.key}" "producer"
+          "span.origin.type" ActiveMQMessageProducer.name
+        }
+      }
+      span(2) {
+        childOf span(1)
+        serviceName "jms"
+        operationName "jms.produce"
+        resourceName "Produced for $jmsResourceName"
+        spanType DDSpanTypes.MESSAGE_PRODUCER
+        errored false
+
+        tags {
+          defaultTags()
+          "${DDTags.SPAN_TYPE}" DDSpanTypes.MESSAGE_PRODUCER
+          "${Tags.COMPONENT.key}" "jms"
+          "${Tags.SPAN_KIND.key}" "producer"
+          "span.origin.type" ActiveMQMessageProducer.name
         }
       }
     }
@@ -282,7 +336,7 @@ class JMS2Test extends AgentTestRunner {
         tags {
           defaultTags()
           "${DDTags.SPAN_TYPE}" DDSpanTypes.MESSAGE_CONSUMER
-          "${Tags.COMPONENT.key}" "jms2"
+          "${Tags.COMPONENT.key}" "jms"
           "${Tags.SPAN_KIND.key}" "consumer"
           "span.origin.type" origin
         }
