+++ settings.gradle
@@ -30,8 +30,7 @@ include ':dd-java-agent:instrumentation:jboss-classloading'
 include ':dd-java-agent:instrumentation:jdbc'
 include ':dd-java-agent:instrumentation:jedis-1.4'
 include ':dd-java-agent:instrumentation:jetty-8'
-include ':dd-java-agent:instrumentation:jms-1'
-include ':dd-java-agent:instrumentation:jms-2'
+include ':dd-java-agent:instrumentation:jms'
 include ':dd-java-agent:instrumentation:jsp-2.3'
 include ':dd-java-agent:instrumentation:kafka-clients-0.11'
 include ':dd-java-agent:instrumentation:kafka-streams-0.11'
